+++ packages/chrome-devtools/src/index.tsx
@@ -3,7 +3,6 @@ import App from './App';
 
 (async () => {
   if (process.env.NODE_ENV === 'development') {
-    // @ts-expect-error
     await import('../mock');
   }
   const container = document.getElementById('root') as HTMLElement;