+++ packages/runtime-core/src/plugins/snapshot/disabled.ts
@@ -1,0 +1,5 @@
+import { PluginSystem } from '../../utils/hooks';
+
+export class DisabledSnapshotHandler {
+  hooks = new PluginSystem({});
+}