+++ packages/storybook-addon/jest.config.cjs
@@ -1,0 +1,34 @@
+/* eslint-disable */
+module.exports = {
+  displayName: 'storybook-addon',
+  preset: '../../jest.preset.js',
+  transform: {
+    '^.+\\.[tj]sx?$': [
+      '@swc/jest',
+      {
+        jsc: {
+          parser: {
+            syntax: 'typescript',
+            tsx: true,
+          },
+          transform: {
+            react: {
+              runtime: 'automatic',
+            },
+          },
+          target: 'es2022',
+        },
+        module: {
+          type: 'commonjs',
+        },
+      },
+    ],
+  },
+  moduleNameMapper: {
+    '^(\\.{1,2}/.*)\\.js$': '$1',
+  },
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'html'],
+  coverageDirectory: '../../coverage/packages/storybook-addon',
+  // https://mswjs.io/docs/migrations/1.x-to-2.x#requestresponsetextencoder-is-not-defined-jest
+  testEnvironment: 'jest-fixed-jsdom',
+};